Increases queryform custom fields from 3 to 8. Allows OR and ().
Currently the occurrence editor search tool (queryform) is limited to three custom fields. This increases the number of available custom fields to eight. It also allows for the use of OR in searches in addition to AND and NOT. To accommodate OR, it also allows for the use of non-nested parentheses for custom fields. These changes greatly improve the flexibility of the occurrence editor search.

This implements improvements to the queryform that were originally implemented by Ben Brandt in Symbiota on 2017-11-27 here: https://github.com/Symbiota/Symbiota/commit/2597b626cf80c72370c18179b1ad522f891f2b23