refactors to add options to limit query and link display
To avoid issues in portals with a large number of records, and to customize display of counts, a few options have been included:

- `getOccTaxonInDbCnt` by default doesn't query the database for counts if taxonrank is above genus (170)
- `getSearchByTaxon` by default will limit showing search urls for taxa with less than 200k records (to avoid ultra slow searches)

Both methods in the current code are being called with optional values (in this case limiting the occurrence counts by taxa above family and limiting url outputs for taxa with more than 500k). Will adjust the code to remove options and use defaults.