Adds a file attachment system to loans and exchanges
This allows a collection editor to save digital files with the correspondence attached to a particular shipment: shipping notices, loan acknowledgements, specimen lists, signed agreements, emails, and any other information relevant to that loan, exchange, or gift. Optionally, users can provide a descriptive title for each attachment.

The maximum attachment size is set at 10 MB, and files are currently restricted to the following filetypes: .pdf, .jpg, .jpeg, .png, .txt, .csv, .xls, .xlsx, .doc, and .docx.

Files are physically saved in content/collections/loans under a subfolder with the collection ID. The server needs to have write access to content/collections/loans. To minimize duplicate filenames, the loan or transaction number and the date are appended to the original filename. Files are also logged in a new database table omoccurloansattachments, the schema of which is in config/schema/dev/db_schema_patch-dev.sql.