Improved autogeneration of next loan/exchange ID
The previous code took the last three loans/exchanges, pulled the number or numbers out of the identifier, and returned the largest number found, incremented by one. However, there was a bug that prevented numbers in some identifiers from being counted at all, by skipping any identifier that had non-numeric characters in front of a number. In addition, taking the highest number found within an identifier with multiple numbers might be incorrect (e.g., for `2022-10` the code would increment 2022, not 10). Finally, any alphanumeric characters would not be included in the new identifier.

This fix solves all of these problems
- It returns a full alphanumeric identifier with the number incremented by one, if the largest recent identifier is alphanumeric
- It uses the *last* number found in a string, which is most likely the one to increment
- It should not skip some identifier formats