Adds first and last buttons to the Occurrence Table View page
These show the first 1000 or last 1000 (or other number of) records, allowing a user to quickly get to the beginning or end of a list of records.

The same functionality is in the editor view, and this was a request during the last Symbiota support group meeting. It was pretty simple to add.