Adds ability to export associated occurrences in new JSON format
This allows associated occurrences to be packaged up as JSON into the associatedOccurrences field. Together with the import of that JSON, it allows for import/export of associated occurrences between portals, complete with all the data for those associations. Currently, this is only enabled for schema = symbiota. For dwc schema, there is a simpler export that cannot be re-imported.

For internal associations where the association is to another record in the same portal, the code checks for cases where for a given specimen it is listed as the occidAssociate in the association. In those cases, the occid and occidAssociate are switched, and the inverse relationship is used. This means that while one association in omoccurassociations links to two specimens, on export, each of those specimens gets their own data for that association. Additionally, for internal associations, the identifier (guid) and referenceUrl are added in case the data is imported to a portal that does not have the other occurrence (in which case the record becomes an external association). Guids to use as the identifier are sourced from the guidoccurrences table.